requirements.txt
asgiref==3.7.2
certifi==2021.10.8
charset-normalizer==2.0.9
dj-database-url==0.5.0
Django==5.0.1
django-browser-reload==1.12.0
django-tailwind==3.8.0
gunicorn==20.1.0
idna==3.3
Pillow==8.4.0
psycopg2==2.9.2
python-dateutil
python-decouple==3.5
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
typing_extensions==4.9.0
urllib3==1.26.7
whitenoise==5.3.0
django-allauthasgiref==3.7.2
cachetools==5.3.2
certifi==2021.10.8
cffi==1.16.0
chardet==5.2.0
charset-normalizer==2.0.9
cryptography==42.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==5.0.1
django-allauth==0.61.0
django-browser-reload==1.12.0
django-nested-admin==4.0.2
django-quill-editor==0.1.40
django-tailwind==3.8.0
et-xmlfile==1.1.0
google-auth==2.28.1
google-auth-oauthlib==1.2.0
gspread==6.0.2
gunicorn==20.1.0
httplib2==0.22.0
idna==3.3
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.2
pillow==10.2.0
psycopg2==2.9.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==1.5
python-decouple==3.5
python-monkey-business==1.0.0
python3-openid==3.2.0
pytz==2021.3
reportlab==4.1.0
requests==2.26.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
StrEnum==0.4.15
typing_extensions==4.9.0
urllib3==1.26.7
whitenoise==5.3.0
dj-database-url
